+++ R/tables.R
@@ -1,5 +1,11 @@
 #' Convenience functions for reading/writing DBMS tables
 #'
+#' @description [dbWriteTable()] executes several SQL statements that
+#' create/overwrite a table and fill it with values.
+#' \pkg{RPostgres} does not use parameterised queries to insert rows because
+#' benchmarks revealed that this was considerably slower than using a single
+#' SQL string.
+#'
 #' @param conn a [PqConnection-class] object, produced by
 #'   [DBI::dbConnect()]
 #' @param name a character string specifying a table name. Names will be
@@ -19,9 +25,6 @@
 #'   all postgres servers (e.g. Amazon's redshift). If `FALSE`, generates
 #'   a single SQL string. This is slower, but always supported.
 #'
-#'   RPostgres does not use parameterised queries to insert rows because
-#'   benchmarks revealed that this was considerably slower than using a single
-#'   SQL string.
 #' @examples
 #' # For running the examples on systems without PostgreSQL connection:
 #' run <- postgresHasDefault()
@@ -65,8 +68,11 @@ setMethod("dbWriteTable", c("PqConnection", "character", "data.frame"),
     if (overwrite && append) {
       stopc("overwrite and append cannot both be TRUE")
     }
+    if (!is.null(field.types) && !(is.character(field.types) && !is.null(names(field.types)) && !anyDuplicated(names(field.types)))) {
+      stopc("`field.types` must be a named character vector with unique names, or NULL")
+    }
     if (append && !is.null(field.types)) {
-      stopc("Cannot specify field.types with append = TRUE")
+      stopc("Cannot specify `field.types` with `append = TRUE`")
     }
 
     found <- dbExistsTable(conn, name)
@@ -78,22 +84,31 @@ setMethod("dbWriteTable", c("PqConnection", "character", "data.frame"),
       dbRemoveTable(conn, name)
     }
 
+    value <- sqlRownamesToColumn(value, row.names)
+
     if (!found || overwrite) {
-      if (!is.null(field.types)) {
-        if (is.null(names(field.types)))
-          types <- structure(field.types, .Names = colnames(value))
-        else
-          types <- field.types
+      if (is.null(field.types)) {
+        combined_field_types <- lapply(value, dbDataType, dbObj = conn)
       } else {
-        types <- value
+        combined_field_types <- rep("", length(value))
+        names(combined_field_types) <- names(value)
+        field_types_idx <- match(names(field.types), names(combined_field_types))
+        stopifnot(!any(is.na(field_types_idx)))
+        combined_field_types[field_types_idx] <- field.types
+        values_idx <- setdiff(seq_along(value), field_types_idx)
+        combined_field_types[values_idx] <- lapply(value[values_idx], dbDataType, dbObj = conn)
       }
-      sql <- sqlCreateTable(conn, name, if (is.null(field.types)) value else types,
-        row.names = row.names, temporary = temporary)
-      dbExecute(conn, sql)
+
+      dbCreateTable(
+        conn = conn,
+        name = name,
+        fields = combined_field_types,
+        temporary = temporary
+      )
     }
 
     if (nrow(value) > 0) {
-      value <- sqlData(conn, value, row.names = row.names, copy = copy)
+      value <- sqlData(conn, value, row.names = FALSE, copy = copy)
       if (!copy) {
         sql <- sqlAppendTable(conn, name, value)
         dbExecute(conn, sql)
@@ -115,8 +130,9 @@ setMethod("dbWriteTable", c("PqConnection", "character", "data.frame"),
 
 #' @export
 #' @inheritParams DBI::sqlRownamesToColumn
+#' @param ... Ignored.
 #' @rdname postgres-tables
-setMethod("sqlData", "PqConnection", function(con, value, row.names = FALSE, copy = TRUE) {
+setMethod("sqlData", "PqConnection", function(con, value, row.names = FALSE, ...) {
   if (is.null(row.names)) row.names <- FALSE
   value <- sqlRownamesToColumn(value, row.names)
 
@@ -166,6 +182,27 @@ format_keep_na <- function(x, ...) {
   ret
 }
 
+#' @description [dbAppendTable()] is overridden because \pkg{RPostgres}
+#' uses placeholders of the form `$1`, `$2` etc. instead of `?`.
+#' @rdname postgres-tables
+#' @export
+setMethod("dbAppendTable", signature("DBIConnection"),
+  function(conn, name, value, ..., row.names = NULL) {
+    stopifnot(is.null(row.names))
+
+    query <- sqlAppendTableTemplate(
+      con = conn,
+      table = name,
+      values = value,
+      row.names = row.names,
+      prefix = "$",
+      pattern = "1",
+      ...
+    )
+    values <- sqlRownamesToColumn(value, row.names)
+    dbExecute(conn, query, param = unname(as.list(value)))
+  }
+)
 
 #' @export
 #' @param check.names If `TRUE`, the default, column names will be
@@ -200,7 +237,7 @@ setMethod("dbListTables", "PqConnection", function(conn, ...) {
   query <- paste0(
     "SELECT table_name FROM INFORMATION_SCHEMA.tables ",
     "WHERE ",
-    "(table_schema = ANY(current_schemas(false)) OR table_type = 'LOCAL TEMPORARY')"
+    "(table_schema = ANY(current_schemas(true))) AND (table_schema <> 'pg_catalog')"
   )
   dbGetQuery(conn, query)[[1]]
 })
@@ -223,37 +260,69 @@ setMethod("dbExistsTable", c("PqConnection", "Id"), function(conn, name, ...) {
 })
 
 exists_table <- function(conn, id) {
-  table <- dbQuoteString(conn, id[["table"]])
-
   query <- paste0(
-    "SELECT COUNT(*) FROM INFORMATION_SCHEMA.tables WHERE table_name = ",
-    table
+    "SELECT COUNT(*) FROM ",
+    find_table(conn, id)
   )
-
-  if ("schema" %in% names(id)) {
-    query <- paste0(
-      query,
-      "AND ",
-      "table_schema = ",
-      dbQuoteString(conn, id[["schema"]])
-    )
-  } else {
-    query <- paste0(
-      query,
-      "AND ",
-      "(table_schema = ANY(current_schemas(false)) OR table_type = 'LOCAL TEMPORARY')"
-    )
-  }
 
   dbGetQuery(conn, query)[[1]] >= 1
 }
 
+find_table <- function(conn, id, inf_table = "tables", only_first = FALSE) {
+  if ("schema" %in% names(id)) {
+    query <- paste0(
+      "(SELECT 1 AS nr, ",
+      dbQuoteString(conn, id[["schema"]]), "::varchar",
+      " AS table_schema) t"
+    )
+  } else {
+    # https://stackoverflow.com/a/8767450/946850
+    query <- paste0(
+      "(SELECT nr, schemas[nr] AS table_schema FROM ",
+      "(SELECT *, generate_subscripts(schemas, 1) AS nr FROM ",
+      "(SELECT current_schemas(true) AS schemas) ",
+      "t) ",
+      "tt WHERE schemas[nr] <> 'pg_catalog') ",
+      "ttt"
+    )
+  }
+
+  table <- dbQuoteString(conn, id[["table"]])
+  query <- paste0(
+    query, " ",
+    "INNER JOIN INFORMATION_SCHEMA.", inf_table, " USING (table_schema) ",
+    "WHERE table_name = ", table
+  )
+
+  if (only_first) {
+    # https://stackoverflow.com/a/31814584/946850
+    query <- paste0(
+      "(SELECT *, rank() OVER (ORDER BY nr) AS rnr ",
+      "FROM ", query,
+      ") tttt WHERE rnr = 1"
+    )
+  }
+
+  query
+}
+
 #' @export
 #' @rdname postgres-tables
+#' @param temporary If `TRUE`, only temporary tables are considered.
+#' @param fail_if_missing If `FALSE`, `dbRemoveTable()` succeeds if the
+#'   table doesn't exist.
 setMethod("dbRemoveTable", c("PqConnection", "character"),
-  function(conn, name, ...) {
+  function(conn, name, ..., temporary = FALSE, fail_if_missing = TRUE) {
     name <- dbQuoteIdentifier(conn, name)
-    dbExecute(conn, paste("DROP TABLE ", name))
+    if (fail_if_missing) {
+      extra <- ""
+    } else {
+      extra <- "IF EXISTS "
+    }
+    if (temporary) {
+      extra <- paste0(extra, "pg_temp.")
+    }
+    dbExecute(conn, paste0("DROP TABLE ", extra, name))
     invisible(TRUE)
   }
 )
@@ -262,11 +331,34 @@ setMethod("dbRemoveTable", c("PqConnection", "character"),
 #' @rdname postgres-tables
 setMethod("dbListFields", c("PqConnection", "character"),
   function(conn, name, ...) {
-    name <- dbQuoteString(conn, name)
-    query <- paste0("SELECT column_name FROM information_schema.columns WHERE table_name = ", name)
-    dbGetQuery(conn, query)[[1]]
+    quoted <- dbQuoteIdentifier(conn, name)
+    id <- dbUnquoteIdentifier(conn, quoted)[[1]]@name
+
+    list_fields(conn, id)
   }
 )
+
+#' @export
+#' @rdname postgres-tables
+setMethod("dbListFields", c("PqConnection", "Id"),
+  function(conn, name, ...) {
+    list_fields(conn, name@name)
+  }
+)
+
+list_fields <- function(conn, id) {
+  query <- find_table(conn, id, "columns", only_first = TRUE)
+  query <- paste0(
+    "SELECT column_name FROM ",
+    query, " ",
+    "ORDER BY ordinal_position"
+  )
+  fields <- dbGetQuery(conn, query)[[1]]
+  if (length(fields) == 0) {
+    stop("Table ", dbQuoteIdentifier(conn, id), " not found.", call. = FALSE)
+  }
+  fields
+}
 
 #' @export
 #' @inheritParams DBI::dbListObjects
@@ -277,7 +369,7 @@ setMethod("dbListObjects", c("PqConnection", "ANY"), function(conn, prefix = NUL
     query <- paste0(
       "SELECT NULL AS schema, table_name AS table FROM INFORMATION_SCHEMA.tables\n",
       "WHERE ",
-      "(table_schema = ANY(current_schemas(false)) OR table_type = 'LOCAL TEMPORARY')\n",
+      "(table_schema = ANY(current_schemas(true))) AND (table_schema <> 'pg_catalog')\n",
       "UNION ALL\n",
       "SELECT DISTINCT table_schema AS schema, NULL AS table FROM INFORMATION_SCHEMA.tables"
     )