+++ NEWS.md
@@ -1,7 +1,22 @@
+# RPostgres 1.1.1 (2018-05-05)
+
+- Add support for `bigint` argument to `dbConnect()`, supported values are `"integer64"`, `"integer"`, `"numeric"` and `"character"`. Large integers are returned as values of that type (r-dbi/DBItest#133).
+- Data frames resulting from a query always have unique non-empty column names (r-dbi/DBItest#137).
+- New arguments `temporary` and `fail_if_missing` (default: `TRUE`) to `dbRemoveTable()` (r-dbi/DBI#141, r-dbi/DBI#197).
+- Using `dbCreateTable()` and `dbAppendTable()` internally (r-dbi/DBI#74).
+- The `field.types` argument to `dbWriteTable()` now must be named.
+- Using `current_schemas(true)` also in `dbListObjects()` and `dbListTables()`, for consistency with `dbListFields()`. Objects from the `pg_catalog` schema are still excluded.
+- `dbListFields()` doesn't list fields from tables found in the `pg_catalog` schema.
+- The `dbListFields()` method now works correctly if the `name` argument is a quoted identifier or of class `Id`, and throws an error if the table is not found (r-dbi/DBI#75).
+- Implement `format()` method for `SqliteConnection` (r-dbi/DBI#163).
+- Reexporting `Id()`, `DBI::dbIsReadOnly()` and `DBI::dbCanConnect()`.
+- Now imports DBI 1.0.0.
+
+
 # RPostgres 1.1.0 (2018-04-04)
 
 - Breaking change: `dbGetException()` is no longer reexported from DBI.
-- Make "typname" information available after `dbFetch()` and `dbGetQuery()`. Values of unknown type are returned as character vector of class `"pq_xxx"`, where `"xxx"` is the "typname" returned from PostgreSQL. In particular, `JSON` and `JSONB` values now have class `"pq_json"` and `"pq_jsonb"`, respectively. The return value of `dbColumnInfo()` gains new columns `".oid"` (`integer`), `". known"` (`logical`) and `".typname"` (`character`)(#114, @etiennebr).
+- Make "typname" information available after `dbFetch()` and `dbGetQuery()`. Values of unknown type are returned as character vector of class `"pq_xxx"`, where `"xxx"` is the "typname" returned from PostgreSQL. In particular, `JSON` and `JSONB` values now have class `"pq_json"` and `"pq_jsonb"`, respectively. The return value of `dbColumnInfo()` gains new columns `".oid"` (`integer`), `". known"` (`logical`) and `".typname"` (`character`) (#114, @etiennebr).
 - Values of class `"integer64"` are now supported for `dbWriteTable()` and `dbBind()` (#178).
 - Schema support, as specified by DBI: `dbListObjects()`, `dbUnquoteIdentifier()` and `Id()`.
 - Names in the `x` argument to `dbQuoteIdentifier()` are preserved in the output (r-lib/DBI#173).